requirements.txt
streamlit
tensorflow
torch
pillow
numpy
